refactor(common-types): relocate memory token schemas out of routes/types (PR-2m step 1)
`PreviewTokenSchema`/`PurgeTokenSchema` were defined in `routes/types.ts` but
their only consumer is `schemas/api/memory.ts`, creating a domain→routes
back-edge (`schemas → routes`). They're branded Zod schemas for memory
preview/purge tokens — domain contract, not route infrastructure.

Move them (and their colocated tests) into `schemas/api/memory.ts`, restoring
the strict one-way `routes → schemas` direction. This is step 1 of the
api-contract / api-client extraction (PR-2m): the back-edge would otherwise
become a circular PACKAGE dependency once routes/ and schemas/ split. Zero
package churn, fully internal — no consumer imports these symbols directly
(verified across services).
