+++ packages/common-types/src/schemas/api/memory.test.ts
@@ -6,6 +6,8 @@
 
 import { describe, it, expect } from 'vitest';
 import {
+  PreviewTokenSchema,
+  PurgeTokenSchema,
   FocusModeSchema,
   SetMemoryLockSchema,
   MemoryUpdateSchema,
@@ -596,6 +598,39 @@ describe('Memory API Input Schema Tests', () => {
 
     it('rejects { success: "true" }', () => {
       expect(DeleteMemoryResponseSchema.safeParse({ success: 'true' }).success).toBe(false);
+    });
+  });
+
+  describe('PreviewTokenSchema', () => {
+    it('accepts a properly-formatted token', () => {
+      expect(PreviewTokenSchema.safeParse('preview_test0000test0000test0000').success).toBe(true);
+    });
+
+    it('rejects an arbitrary string', () => {
+      expect(PreviewTokenSchema.safeParse('not-a-token').success).toBe(false);
+      expect(PreviewTokenSchema.safeParse('').success).toBe(false);
+    });
+
+    it('rejects a string lacking the preview_ prefix', () => {
+      expect(PreviewTokenSchema.safeParse('purge_test0000test0000test0000').success).toBe(false);
+    });
+
+    it('rejects a token that is too short (no payload)', () => {
+      expect(PreviewTokenSchema.safeParse('preview_x').success).toBe(false);
+    });
+  });
+
+  describe('PurgeTokenSchema', () => {
+    it('accepts a properly-formatted token', () => {
+      expect(PurgeTokenSchema.safeParse('purge_test0000test0000test0000').success).toBe(true);
+    });
+
+    it('rejects a preview-prefixed token (distinct brand)', () => {
+      expect(PurgeTokenSchema.safeParse('preview_test0000test0000test0000').success).toBe(false);
+    });
+
+    it('rejects unbranded raw strings', () => {
+      expect(PurgeTokenSchema.safeParse('test0000test0000test0000').success).toBe(false);
     });
   });
 });
